requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
pydantic==2.5.0
pyjwt==2.12.1
passlib==1.7.4
python-multipart==0.0.6
psycopg2-binary==2.9.9
supabase==2.4.0